Converts an acceleration in the aggregated file
Added parser of the file aggregating all the data and from it a group
extracted that is related to left thigh experiments. It creates a new
file based on the original one with 'linearized' prefix where
acceleration value is linear (i.e. with g-vector extracted from it).

The format is as follows:
0-based timestamp, lin acc X, lin acc Y, .., Z, gyro X, .. Y, .. Z